+++ sos21-database/src/query/list_registration_forms_by_pending_project.rs
@@ -7,10 +7,16 @@ use anyhow::{Context, Result};
 use futures::stream::{BoxStream, StreamExt};
 use uuid::Uuid;
 
+#[derive(Debug, Clone)]
+pub struct PendingProjectRegistrationFormData {
+    pub registration_form: RegistrationFormData,
+    pub has_answer: bool,
+}
+
 pub fn list_registration_forms_by_pending_project<'a, 'b, E>(
     conn: E,
     pending_project_id: Uuid,
-) -> BoxStream<'b, Result<RegistrationFormData>>
+) -> BoxStream<'b, Result<PendingProjectRegistrationFormData>>
 where
     E: sqlx::Executor<'a, Database = sqlx::Postgres> + 'b,
     'a: 'b,
@@ -38,12 +44,15 @@ SELECT
         ))
         /* works because attributes column in registration_form_project_query_conjunctions table is NOT NULL */
         FILTER (WHERE registration_form_project_query_conjunctions.attributes IS NOT NULL)
-        AS "query: Vec<(Option<ProjectCategory>, ProjectAttributes)>"
+        AS "query: Vec<(Option<ProjectCategory>, ProjectAttributes)>",
+    bool_or(registration_form_answers.id IS NOT NULL) AS has_answer
 FROM pending_project_registration_forms
 INNER JOIN registration_forms
     ON registration_forms.id = pending_project_registration_forms.id
 LEFT OUTER JOIN registration_form_project_query_conjunctions
     ON registration_forms.id = registration_form_project_query_conjunctions.registration_form_id
+LEFT OUTER JOIN registration_form_answers
+    ON registration_forms.id = registration_form_answers.registration_form_id AND registration_form_answers.pending_project_id = $1
 GROUP BY registration_forms.id
 "#,
         pending_project_id
@@ -70,9 +79,12 @@ GROUP BY registration_forms.id
             })
             .collect();
 
-        Ok(RegistrationFormData {
-            registration_form,
-            query,
+        Ok(PendingProjectRegistrationFormData {
+            registration_form: RegistrationFormData {
+                registration_form,
+                query,
+            },
+            has_answer: row.has_answer.unwrap_or(false)
         })
     })
     .boxed()